Episode 5B - Custom Configs
Showed how to create a custom config file called "data.yml." Also showed how to read and set data in the file. Created a command to send a message with the information, as well as another command to reload the config instead of reloading the server.

Also changed the name of the previous episode folder the 5A since I forgot to do that when I uploaded it.